swagger docs first build
